Restructures folders issue, status and compliance queries now in separate files
